+++ src/Page.php
@@ -87,6 +87,11 @@ class Page {
             $tab_obj->set_name( $tab );
         }
 
+        // if position is already used, add + 1.
+        if( isset( $this->tabs[$position]) ) {
+            $position++;
+        }
+
         // add the tab to the list of tabs of these settings.
         $this->tabs[$position] = $tab_obj;